requirements.txt
Django>=5.1
djangorestframework>=3.15
django-cors-headers>=4.4
Pillow>=11.0.0
djangorestframework-simplejwt>=5.4
channels>=4.1
channels-redis>=4.2
redis>=5.0.1
django-extensions>=3.2.3
